+++ osu.Game.Rulesets.Karaoke/Skinning/Groups/StyleMappingBySingerIds.cs
@@ -4,7 +4,7 @@
 using Newtonsoft.Json;
 using osu.Game.Rulesets.Karaoke.Skinning.Elements;
 
-namespace osu.Game.Rulesets.Karaoke.Skinning.Mappings
+namespace osu.Game.Rulesets.Karaoke.Skinning.Groups
 {
     public class StyleMappingBySingerIds : IStyleMappingRole
     {
