+++ osu.Game.Rulesets.Karaoke/Skinning/Groups/StyleMappingBySingerNumber.cs
@@ -4,7 +4,7 @@
 using Newtonsoft.Json;
 using osu.Game.Rulesets.Karaoke.Skinning.Elements;
 
-namespace osu.Game.Rulesets.Karaoke.Skinning.Mappings
+namespace osu.Game.Rulesets.Karaoke.Skinning.Groups
 {
     public struct StyleMappingBySingerNumber : IStyleMappingRole
     {
